Rename files to keep the patterns
